ml_internship_project/requirements.txt
accelerate==0.28.0
bitsandbytes==0.43.0
bcrypt==4.1.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
langchain==0.1.12
langchain-community==0.0.28
langchain-core==0.1.31
langchain-openai==0.0.8
langchain-community==0.0.28
langchain-text-splitters==0.0.1
PyPDF2==2.11.1
python-dotenv==0.21.0
pysqlite3-binary==0.5.2.post3
streamlit==1.32.1